Manage deletion of uploaded files
Transfer remove method from VoyagerMediaController to VoyagerBaseController
Created new route
changed permission check
remove file and thumbnails from file system
