+++ resources/views/bread/edit-add.blade.php
@@ -183,7 +183,7 @@
             $('.form-group').on('click', '.remove-single-file', deleteHandler('a', false));
 
             $('#confirm_delete').on('click', function(){
-                $.post('{{ route('voyager.media.remove') }}', params, function (response) {
+                $.post('{{ route('voyager.'.$dataType->slug.'.remove') }}', params, function (response) {
                     if ( response
                         && response.data
                         && response.data.status
