+++ src/Http/Controllers/VoyagerBaseController.php
@@ -2,6 +2,7 @@
 
 namespace TCG\Voyager\Http\Controllers;
 
+use Exception;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
@@ -464,6 +465,138 @@ class VoyagerBaseController extends Controller
         }
 
         return redirect()->route("voyager.{$dataType->slug}.index")->with($data);
+    }
+
+    //***************************************
+    //
+    //  Delete uploaded file
+    //
+    //****************************************
+
+    public function remove(Request $request)
+    {
+        try {
+            // GET THE SLUG, ex. 'posts', 'pages', etc.
+            $slug = $request->get('slug');
+
+            // GET file name
+            $filename = $request->get('filename');
+
+            // GET record id
+            $id = $request->get('id');
+
+            // GET field name
+            $field = $request->get('field');
+
+            // GET multi value
+            $multi = $request->get('multi');
+
+            $dataType = Voyager::model('DataType')->where('slug', '=', $slug)->first();
+
+            // Load model and find record
+            $model = app($dataType->model_name);
+            $data = $model::find([$id])->first();
+
+            // Check if field exists
+            if (!isset($data->{$field})) {
+                throw new Exception(__('voyager::generic.field_does_not_exist'), 400);
+            }
+
+            // Check permission
+            $this->authorize('edit', $data);
+
+            if (@json_decode($multi)) {
+                // Check if valid json
+                if (is_null(@json_decode($data->{$field}))) {
+                    throw new Exception(__('voyager::json.invalid'), 500);
+                }
+
+                // Decode field value
+                $fieldData = @json_decode($data->{$field}, true);
+                $key = null;
+
+                // Check if we're dealing with a nested array for the case of multiple files
+                if (is_array($fieldData[0])) {
+                    foreach ($fieldData as $index=>$file) {
+                        $file = array_flip($file);
+                        if (array_key_exists($filename, $file)) {
+                            $key = $index;
+                            break;
+                        }
+                    }
+                } else {
+                    $key = array_search($filename, $fieldData);
+                }
+
+                // Check if file was found in array
+                if (is_null($key) || $key === false) {
+                    throw new Exception(__('voyager::media.file_does_not_exist'), 400);
+                }
+
+                $fileToRemove = $fieldData[$key];
+
+                // Remove file from array
+                unset($fieldData[$key]);
+
+                // Generate json and update field
+                $data->{$field} = empty($fieldData) ? null : json_encode(array_values($fieldData));
+            } else {
+                if ($filename == $data->{$field}) {
+                    $fileToRemove = $data->{$field};
+
+                    $data->{$field} = null;
+                } else {
+                    throw new Exception(__('voyager::media.file_does_not_exist'), 400);
+                }
+            }
+
+            // Remove file from filesystem
+            if ($fileToRemove != config('voyager.user.default_avatar')) {
+                $this->deleteFileIfExists($fileToRemove);
+            }
+
+            $row = $dataType->rows->where('field', $field)->first();
+
+            if (!empty($row->details->thumbnails)) {
+                $ext = explode('.', $fileToRemove);
+                $extension = '.'.$ext[count($ext) - 1];
+
+                $path = str_replace($extension, '', $fileToRemove);
+
+                foreach ($row->details->thumbnails as $thumbnail) {
+                    $thumb_name = $thumbnail->name;
+
+                    $this->deleteFileIfExists($path.'-'.$thumb_name.$extension);
+                }
+            }
+
+            $data->save();
+
+            return response()->json([
+               'data' => [
+                   'status'  => 200,
+                   'message' => __('voyager::media.file_removed'),
+               ],
+            ]);
+        } catch (Exception $e) {
+            $code = 500;
+            $message = __('voyager::generic.internal_error');
+
+            if ($e->getCode()) {
+                $code = $e->getCode();
+            }
+
+            if ($e->getMessage()) {
+                $message = $e->getMessage();
+            }
+
+            return response()->json([
+                'data' => [
+                    'status'  => $code,
+                    'message' => $message,
+                ],
+            ], $code);
+        }
     }
 
     /**
